backend/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
Deprecated==1.2.13
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
pycodestyle==2.7.0
pycparser==2.20
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.4.0
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
wrapt==1.13.2
